requirements.txt
fastapi
uvicorn
sqlalchemy
alembic
databases
pydantic
psycopg2-binary
python-dotenv
asyncpg
